Add files via upload
Here, I am not able to connect to the The Things Network (TTN) Console. I have registered an gateway and created an end device also. Earlier it has given the payload data but I am not getting dht11 sensor data on TTN when I did next time. It is giving EV_Join Failed message and I am not able to connect to TTN. Could you please check the codes and notify me where the problem lies.